TINY-10209: Fix fullscreen toggle not working
