+++ modules/oxide/src/less/theme/components/dialog/dialog.less
@@ -385,11 +385,6 @@
     max-width: 1200px;
   }
 
-  .tox-dialog--width-lg--inline {
-    height: 100%; // The height property is required to flex the textarea within large dialogs to fit to parent
-    max-width: 1200px;
-  }
-
   .tox-dialog--fullscreen {
     height: 100%;
     max-width: 100%;
